move get_configured_variables to config related things
